target/classes/top/chenqwwq/leetcode/daily/_20200825/Solution_test.go
package _20200825

import (
	"reflect"
	"testing"
)

/**
  @user: chenqwwq
  @date: 2020/8/25
*/

func Test_findSubsequences(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		/**
		示例:

		输入: [4, 6, 7, 7]
		输出: [[4, 6], [4, 7], [4, 6, 7], [4, 6, 7, 7], [6, 7], [6, 7, 7], [7,7], [4,7,7]]
		说明:
		*/
		{
			name: "1",
			args: args{
				nums: []int{4, 6, 7, 7},
			},
			want: [][]int{{4, 6}, {4, 7}, {4, 6, 7}, {4, 6, 7, 7}, {6, 7}, {6, 7, 7}, {7, 7}, {4, 7, 7}},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findSubsequences(tt.args.nums); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("findSubsequences() = %v, want %v", got, tt.want)
			}
		})
	}
}
